+++ lib/src/styled/styled.dart
@@ -1,0 +1,18 @@
+/*
+
+final Styled style = Styled.widget()
+
+
+
+
+*/
+
+import 'package:flutter/semantics.dart';
+import 'package:flutter/widgets.dart';
+
+class Styled {
+ 
+}
+
+
+