+++ lib/src/styled/styledIcon.dart
@@ -1,0 +1,5 @@
+import 'styled.dart';
+
+class StyledIcon extends Styled {
+
+}