+++ lib/src/styled/styledText.dart
@@ -1,0 +1,5 @@
+import 'styled.dart';
+
+class StyledText extends Styled {
+
+}