+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## [0.0.2]
+* Added more basic widgets
+* Added some animated widgets
+
 ## [0.0.1] - TODO: Add release date.
 * Initial consept